example-integration cleanup starter package
