+++ digiwf-integrations/digiwf-example-integration/digiwf-example-integration-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,3 @@
 de.muenchen.oss.digiwf.example.integration.ExampleIntegrationAutoconfiguration
+de.muenchen.oss.digiwf.example.integration.ExampleIntegrationAdapterInAutoconfiguration
+de.muenchen.oss.digiwf.example.integration.ExampleIntegrationAdapterOutAutoconfiguration
